Add AudioTempoFilter unit test

+++ FFmpegDSLAndroid/src/main/java/io/github/vnicius/ffmpegdslandroid/ffmpegcommand/filter/AudioTempoFilter.kt
@@ -1,6 +1,6 @@
 package io.github.vnicius.ffmpegdslandroid.ffmpegcommand.filter
 
-class AudioTempoFilter(tempo: Float): Filter() {
+class AudioTempoFilter(tempo: Double): Filter() {
     override val key: String = "atempo"
     override val value: String = tempo.toString()
 }
